requirements.txt
altair==6.1.0
anyio==4.13.0
asttokens==3.0.1
attrs==26.1.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==7.0.6
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
comm==0.2.3
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
filelock==3.29.0
gdown==6.0.0
gitdb==4.0.12
GitPython==3.1.49
h11==0.16.0
httptools==0.7.1
idna==3.13
ipykernel==7.2.0
ipython==9.13.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
narwhals==2.20.0
nest-asyncio==1.6.0
numpy==2.4.4
packaging==26.2
pandas==3.0.2
parso==0.8.6
pillow==12.2.0
platformdirs==4.9.6
prompt_toolkit==3.0.52
protobuf==7.34.1
psutil==7.2.2
pure_eval==0.2.3
pyarrow==24.0.0
pydeck==0.9.2
Pygments==2.20.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.27
pyzmq==27.1.0
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
smmap==5.0.3
soupsieve==2.8.3
stack-data==0.6.3
starlette==1.0.0
streamlit==1.57.0
tenacity==9.1.4
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.6.3
uvicorn==0.46.0
watchdog==6.0.0
wcwidth==0.6.0
websockets==16.0
